LPS-174464 Internalize partialMatchMap into JSModulesCache
With this commit, all information regarding modules resolution is inside the
JSModulesCache.

Now we will extract JSModulesCache to its own class and decouple it from
NPMRegistry so that we can resolve modules without needing to access the
NPMRegistry several times.

This will prevent synchronization issues in the middle of module resolutions if
the NPMRegistry is updated.