LPS-181170 Publish clay packages through import maps
We obtain the list of @clayui packages from package.json. This is so that we
don't need to manually update the list of modules.

Note that we can do this because the contract between DXP and client
extensions is that all packages from the internal version of Clay DXP is
using are exported through import maps.

Thus, any change to this list is expected and legit, even though it may
potentially break customer client extensions, because it doesn't break the
contract itself.