LPS-181170 Add a bash script to format the liferay-sample-workspace
Not sure if this is the best place, but we cannot add it to the workspace's
build.gradle files because that would make customers use
@liferay/npm-scripts, which is something we have never contemplated and can
lead to a lot of problems.

So, while we find a better way to format these repos, this may be a valid
workaround, I guess.