LPS-185489 Skipped flaky test
This test has already failed a few times when updating Clay components but it is unrelated to our changes because there is no Clay component being used here.

In the last rounds of testing it is passing the CI but failing the DXP release, I am skipping this test so the team can investigate the issue later.